updated prev act totals table file name
